o Committing new loadable module design.   - Incorporated Chris Dunlap's list.[ch]. Changed source acc'dngly   - Added mod.[ch] and mod_rcmd.[ch] to handle module loading and usage   - Added modules subdirectory   - Added libtool's ltdl.[ch] for portable dlopen()   - Removed sshcmd.c xrcmd.c and qcmd.c and added modules w/ same functionality   - Rewrote configure.ac, abstracted tests into auxdir/*.m4

+++ xmalloc.c
@@ -62,7 +62,7 @@ void *Malloc(size_t size)
     p = (int *) malloc(size + 2 * sizeof(int));
     MALLOC_UNLOCK();
     if (!p) {
-        fprintf(stderr, "Malloc(%d) failed\n", size);
+        fprintf(stderr, "Malloc(%ld) failed\n", (long) size);
         exit(1);
     }
     p[0] = XMALLOC_MAGIC;       /* add "secret" magic cookie */
@@ -91,7 +91,7 @@ void Realloc(void **item, size_t newsize)
     p = (int *) realloc(p, newsize + 2 * sizeof(int));
     MALLOC_UNLOCK();
     if (!p) {
-        fprintf(stderr, "Realloc(%d) failed\n", newsize);
+        fprintf(stderr, "Realloc(%ld) failed\n", (long) newsize);
         exit(1);
     }
     assert(p[0] == XMALLOC_MAGIC);
